Heat loss for flats from roofs and floors
Flats which have a floor which is the top floor of their building have a heat loss roof.

Flats which have a ground or basement floor have a heat loss floor.

Roof and floor measures now check for the presences of external roof or
floor, instead of whether it's a flat.

Ticket 86 http://help.deccnhm.org.uk/scp/tickets.php?id=91